refactor: remove toolbar action compatibility exports
